Reduce some memory growth issues
If there are lots of transaction (currently testing with 500) causes eventual process death.

It seems that it is because of exponential RAM usage. This is just the first iteration of a fix; the true fix will hopefully come when running with heaptrack and seeing what components are eating the RAM