Add a script that will generate a sequence from the largest respective config_key value the running system can handle
This script executes the proof_type executable till it finds the max
config_key the system supports without killing the process.It also
fine-tunes the value to get as accurate results as it can.

Once the max value is determined, it uses rounded logarithmic
sequencing to reverse-engineer a sequence we can use for the benchmark
graphs. It will also save these sequences to a file.